Prepare for release
Updated Readme, chart, and aded comments.
