+++ Resources/CoreDataCandy-architecture.drawio
@@ -1,1 +1,1 @@
-<mxfile host="Electron" modified="2020-10-26T23:32:54.515Z" agent="5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/13.7.9 Chrome/85.0.4183.121 Electron/10.1.3 Safari/537.36" etag="VCSZ_xgg1nvDiJBIy0ng" version="13.7.9" type="device"><diagram id="w-XJ28GW39ViOU8AllJJ" name="Page-1">5VxZd9u2Ev41Ord5iA83bY+OlzbnpklOfJ3bPvVAJCSioQgGBG2pv77AAOAK2UwsibL9EIcAsXEw881gZqCRf7He/MpQFv9OI5yMPCfajPzLkef5Yy8Q/8marapx/YmjalaMRLquqrgh/2BdaZoVJMJ5oyGnNOEka1aGNE1xyBt1iDF632y2pElz1gytcKfiJkRJt/b/JOKxrnUn8+rFb5isYj31zJuqF2tkGusvyWMU0ftalX818i8YpVw9rTcXOJHUM3RR/a53vC0XxnDK+3T4wObpP9nXz24Wf/94e/X1Ov5w+das9g4lhf5ivVq+NSTAkaCILlLGY7qiKUquqtp3jBZphOU8jihVbT5QmolKV1T+jTnf6u1FBaeiKubrRL9Vc8qJdn6crsppwUL8wBcZJkFshflDX15ugWBeTNeYs63ox3CCOLlrrgNpJlqV7So6iwdN6h8gu9uh+gVl+BJx1KE+WQOH1qmlqy7JeiVmT8hC/A0Tkv2FGJePdJ0VHLNcPMshFyjHf7nebCP+nWXpSs+NGcebh8ndJY8R7Jlmai3WmovuKwkxDeK6bDgHImfw0pjY68nE/pBM7HWo/vHmd5QK3ow+Lf6WYNzehIrEkl73MeH4JkNAiXuhP2zkfBqbTpps6k66jOp6Fk6dHIpTXxzc+j05dTIo3DqvlezzIcnuW6g+ScRy3y2p+GBvsuJAFVW3aOzH5HtBTcu3OVD0XDRwvWxTvTRDGC13lXLCt+WArD2F+IqFqfsN7EpBwhiLvx3kclLECyZfoTSSu7bJaA4daJooEn4vCMPy3bJIQ05ompetM0YzAVxEdnhgMU0ytBjyEXhsYuk+wNIbN8DSdyxgObeAZYmg+5dbGwc9a7md9JRbd1jzdPJq6T4e1KJ6tXTfhU3HobuN7KVSaqoY7eLopWFoEoFKEP/aGqrSE3cNPbRGWUbkEU2UxT6uYvF0TbAeCPp8AZIIbROTTLaj4s9C9i1y0EYkLUcTg4kF3b4/Ad3SNMQDiyHuOTbd4h9Kt3izIURL0Itt/6gX/pSDnY1N8XKjB1elrS7tUSRnfaHwiRIJXc8ZQ9tag4ySlOe1kT/LiopRpm7QZJRpa6PVgNW2lyv7eU6YPS7915iHMVoIHrDIeRcNPjPKaUgTYHAuZTillcm4xkACJbl0ucRMbi3Kt6ZOCLK0KuWcdevyTBTeURhuB5y0QEo6RZeUrc24hBt4CBFMAavqjGXDIGQZnciPSDGOAHao/Ioc8yIT65ygtYSUdJFnNtP22H6AieM2uMr1vS78GOV7HPhxuy7AZ67q5z1xZTakpvdtwv4qyO5Ph6T7/AcwtnmKfxhpLxTClXBaA+qL3TjZPe3/Uj/OA5QB/AoYfCNh9zxJZCRJT7JG37DGxI835RAKtWsLaPgFduuBwbExOGsd/F0LOrpzCzq6B0NH73F+Oc+yhIRgCmsjtw/DvBWj3ipHTkOflZsrzBHMEOyo3jMTmpGMYB/wksKAKZW9vqX0HpTtf/JaZzDSBT9FoFYRN/pTGu1nDy5TmfK7+IouLZoZPuR7gWGfVf9IrwEU/6OztQe8oOsFSeWbrFgkJI/BZhF7HdOotFoyMQeGL8+BpaFkm+q/qRImTV+kLY17aZlgxkD8yFK/USxgrA2gsKgici459X2M4bzDttpkoTA3ZVgbPAOLlzf1z6ZN+ZpMLYefuWeksCFh44NJ2IuLRLjjnppwUJ/4+DRilTiNzmV2hCiGCcpzEqrKa5IMuTXjQQOaZpm1vbGeeepYXB2RFOxhY2kYYAI8LWuJcQh1rJ02DyQJyXJ8FIwaz2ZtjAqmsy5GjR0rRjlTU73/wP6LCwC4fe31wBtSFIKuKLwawgeDYlCfk5J0CPczdy+xgHdmTis1zw4dmQNM5yjUPDS9kQ05Z2RRgHVXvoZVfJXLfLPDgL0ohCG2rgw2sNQfbMgZSnPtspIRVN3HScg3Y5Z2nEtX3mg2Gc0l0W7fv4fUqN1Gn3MUQH1bnowMnLoWk8+1+bvdQ0GpZ8uBaDFWM7RwPP5izZDGLw2t+6bafwpHEH0GT7fqjK+etDc11Qd/OEPU36jnwfnCb/LF2BYH8S18MT8YW7w4DVumCT8ac3Tsm3XYCIfXgoaJlviDhjjMV9R2uXHerwQ356wIZeZL1zN1TJO0IyhO1yflTqz26MFy97wTObptCJcRRB00FKU/9WDyuYofysIBwodPjtT3E4tgPm0OoRame1U72RnIdXaEms1AChc6Az1Vjs08B5VjE7d6QWjdO34xaIaI18csl0f6/6nv6GU8mQQRhcEJXZHQmCt5hkOyVH4EcBbnMS2k0a9SPVRsGDIiVeclYTkHauSFmAVsIGl+hTRJcKitaPAVqxa5atLsnII7Vjc4Ozsb3F5ynU4Ct0UN2K4auM6hDCbfFri1ccIXvRV9OOF8IfQuCnmXG7QbiWGhk1OlomHzwfnd2EttIyvL+46SyBTuKQOiogUt5DYjOSDScAeGsuSUXAwOHi2aSXZBiQx3fCp4Bn2occfXmPwE+MObnx5/7M56LhnghrIfhAh1lJL9LnEeMpJx2LiHkr9qkSeTSiIDjzjnsLMy7wQclignEM3JxeDaBgTwKWNrQ29yGwR8t4xYDrjNLy5V0w/6auJBc2R9mxXekq9P4Dqi7DPDkWRjKQ6l30j9BQFZCzqRnKa1dn2F0q62tetKSWs5qiUhqlnuCmNWW/mSJByzZyGawdhyZ+DIgmnzXHf8XKsiQexqI8ic54KSnxEXJE77bf8NWWegKdVZ2ex9rLgCgSTA+HJDyymAsHoWodUJADIyeSUyWg5shDccp7q94GEVaOcx6Gx4FylGuOGSI05CBbsnxwLzIaC4TO0ts3ePntrrT3tCuPfUGPhPHZnn4yZaeE3Xl+UoHzzY4TBn7MDmKH/emr0vW/iDht8Ddwg6Pyex9Z8amtyL2I5dZ3RwKfS7+Umn7bGeTVunFN+iB4/ssQ5s7sKWLfS1E5t9xFVRN3ykzaN2AbVNYZOkN2qm8ZWOBGPJjqpswyqiLAfZZicQyPXHJxfIDV5c5l7Q9+QZDPorAsGwN8aG00+998fbh3rq6B+/ZQW67dDQviI6rZijN5/Uh3u0vVnXYa3THu4PG572gfYvWAA7qXuhJQiXGYsq68gJ4Swq9K9B+jLXo+ZxRHeIJCZF0obkvW+gOntCctedtu45eBYN7Xm2DEf3cBmO40HAu5b+m8qcmXru777B46mY3VNYnX6gsDc57HFn/IKmkmfJD98b7dwOjfASQTDJAd+TvDlkUuOgXahmMqaVNs+uIbW7LqFaOiF2ULu6VBfcKtXv9CTWdnHi6AI7Hlhgj5CvP+7780JP/pmSn1LU7UviUw2iB1W84x5xPQiO3pp7ULWrUr2Ev3azqnYzKiZRBCHgjze1sLKAFs5kbJ8Z8cZruN4txfhd/VJb4x5CpC2Dkbl9lul7VDvPWUNdQpxYflNwb5cQRbH63U3FINXPl/pX/wI=</diagram></mxfile>
+<mxfile host="Electron" modified="2020-11-19T21:01:11.319Z" agent="5.0 (Macintosh; Intel Mac OS X 10_16_0) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/13.9.9 Chrome/85.0.4183.121 Electron/10.1.5 Safari/537.36" etag="z8vCEQVL4S4aBBv7GP7A" version="13.9.9" type="device"><diagram id="w-XJ28GW39ViOU8AllJJ" name="Page-1">5Vxbc9u2Ev41mhM/xMOLREmPtmy3mZO6nug4p33qQCQkoiEJBgRtqb++uPIKSXStC5U+xCFBcAEudr9d7C40cGfx+icC0vAXHMBo4FjBeuDeDRzHtcc2+4+3bGSLbQ8nsmVFUKDayoY5+guqRku15iiAWa0jxTiiKK03+jhJoE9rbYAQ/FrvtsRRfdQUrGCrYe6DqN36fxTQUH+GNy0f/AzRKlRDT5yxfBAD3Vl9SRaCAL9Wmtz7gTsjGFN5Fa9nMOLc03yR7z1seVpMjMCEdnnhM5kmf6Vfn+w0/P74fP/1Ifx891HP9gVEufpiNVu60SyAAeOIusWEhniFExDdl623BOdJAPk4Frsr+3zGOGWNNmv8E1K6UcsLcopZU0jjSD1tf4v6vAznxIc7PkDLBCArSHd9qOzHv6UygOLUTxDHkJIN60BgBCh6qa8+UEK0KvqVfGYXitVvYLvd4voME3gHKGhxH8VCQqvcUk13KF6x0SO0YH/9CKV/AEL5JY7TnEKSsWtOcgEy+IftTNbs33WarLYy+wUSCtc72aMVe6KEWqm1kqLXUkN0h7CqG9aR2Dm8cCF2Ogqx2yshdlpcf5z/AhImm8Gviz85GDcXoWQx59driCicp0Cw5pXZj47sfIOYenUxtb22oNqOQVK9Y0nqpcOt21FSvV5JqnYmfni2T3vFdtfAdS9i879dYsYBx1tRwRXZtqith/c9x7rnx0xw9IZ1sJ10XT7UJLSVu08oopuCIGkOwb5iodt+Fn4l42kI2d8WclkJoDnhj0AScJauU5yJF3ASSRZ+zxGB/NkyT3yKcJIVvVOCU4ZSiL+wYzJ1NjQEcg881rH0EGDpjGpg6VoGsJwawLJA0MPrrUmCLklvvY56a/fMPfV+DL4fjp3q1SeMuMZqlRk7df/Ca2qCXHj1VrkoN4SATaVbyjtk3ccZjazGGkuK5YoX3/gO9+4HEYL9W0O7V8pnYnthIev2TsVbOpk7HAXCPrF/TXNZGq2XmlGMQZoivl9k92wdVyG7ekBQERLvfBEsYaYvRCnvh9mfBX83z4RpRElBjRFjE3r+1ANDV9emoWFX4FgmQ+cey9A5k3OoFmMh2fxWvfmdE7se6du7tSIu7zbq7p+r5KSrPTwKfr8Zdu1hXVDGx0fdyX7tf4DUD8GCyYBBz9to8EQwxT6OhIBTrsMJLv3XGAoWSM3FyyUknO8g2+g2psjcxeVjVl3da3ZziwW5LXDSACkeoV1iEmu6iGp48IEYQsyqRcuEQcBAHfGPSCAMBOxg/hUZpHnK5umBmENKsshSk5996qCEZ9k1qbJdpw0/2vieBn7sdjzyskz9tCOuTHpl6V2Tsv+IbHfHveL79A0YWw8p7EbamUS4Ak4rQD3bjpPt0MOHamxBQJmAXwaDVxx2b6KIp7XUIDH4BhUmPs4LEhK1KxOoBSm224GzY+PwuhGFsA3oaE8N6GgfDR2d/fJyk6YR8oUrrJzcLgLzkVF9llGlmj0rFpe5I5AAsaJqzXSeiAuCmeAdFgQTzN/6luBXYWz/k1VeFk46k6dAmFVAtf3kTvv1zmlKV36bXOGlwTKLD/meQ7HO8v1AzUEY/r2jNQnOcLxACX+S5osIZaHwWdhahzgovJaUjQHFl2dCpMWdaaj/JlKZFH+B8jReuWcCCRHqh5bqiRQB7W0IDrMmxMfiQ7+GUOx3yEa5LFiMjQlUDs+Z1csZu9fjun55Y8PmZ+poLaxp2OhoGtbztMjp4+7mDcmoGQdqeoTSYr873tQax26UFuzZKDX6H2enNLz04LDd1X0a9is6PBz9axg/7BXj7S6OK4/PdfM+7qAfAaKdx8pGGxf+ZMszrfuwV7wjpQQtcmFsi8diFl/5NK+2+BOznNnFuLSfwnHa2ZESkGQqgsCza+odK0LftJfQ2uvfO4OJN5hypj1/+iTKZrbbYOskNvhj4ajqqJJtsMC2Kfxoj44kWI4pP94QrHqk93TyReoR5g81b/CqXH8sPEK1JUo2csslr1RwK1H7MOHSVZ/I67PLhdswuqawtGuQi+nRxOLSLWxn/00rwJkDzk4DGjyl8cfN87WDf7XtV6m4GSW5z6si2oECNg5KM3gSTWkpitUOEdieZdrCHK2uy+lHCSJcI8oTOiqHw+5+V8T4dZnO4Tfvz+YcPnHaTS2G03GdxJZtT4uQbW3J/B04X9+asHeCzJFrCuKbwppfYJZHtJsNv1kwpQc+LeEgwivkD4pYB4EMEBKJDyJKKQIhRAyRSXurDLQ0+y8YFUGaV0yEeIAFzjn3ACcIFK+FleZeQsaIi0gTTrkfACIe+vo1p6l4B+vQjPi0/4n1uz67MbetcbPy1GpjlKlG2raOZc3d7eV4hQDMGebcwcwnKKV8AbqISLWy4HHepLCzJKASj9QJRh6OhhkVa8yzkZwyBBkSMb6MEVemKIU+WhYR17Mvd6PQ2LWLOPYZF/zSC3jcYUf3zR0d2hC9j/EmZ6ChaU8EBlx8DVHpbeU7LQgWqqV2xFL7Cqpm/UnLQa0liigkF6FNw5GhEvXEumSKebV2yKucbXjv14zNWcY4+QQoY3HSDUbnKE6FmZNetl7eUC48EMIr6PMFLYYQvFajMJOMBIYCnSDkaQ8hKXBNYaL6E0BlxoSGwuCKZ4EUhDnlEtED++lM7d6JwPTC0HPcFT37UW01HdW13rHc7U6zwbVvgIZTmuA93v3BEhIXf8qks8T06yDa0D4Hn4vqyaJA8tjVk93X5+B5i4No9EnSgm47l/xF7ilucxQF3PKdMXY1aewLXddg104cu9LpxV2+zddWlmZP3KDqyHAfRvoXoOm96uqJQb2+otjVa890UJaBlLklTmST9iCl4456l9IZ9rykYi/YDbtu/ob9Oms6PG8p/8mMUef1cY5ii1rGxh02Im7NIPGhYruN7IMz9ark9vbX8zpubUyHCIQJT7tA+xdmSgmqhoQ5COsIoKo/sHyxt4xBcSqoyPpWgn7gBaBI1REakbzz0aCdsvym5Py4UYDqGCy045jK48pXDw4qo7OAN0yCG/4TOuw24dlz0fKA+NTfCx4nOqDZUlarGygcTA87HOab4UCqwIO6muGESyx68wmf1jmeAC6BSPU8+LqKRkSZeLG3Lp8RL/hyyEJD1YxkEqZSVi4636IEiMVRdcj901lTTevJVXZ0ZpX1I5BlzHOuau37XLJR15+hOPxx9n9kqptlqWMFo0c1vaMOaTaRq3zWJeqVKvZO2l4peq8UrYcoCERG9nFeyfIyLKEER5EM+HLthbE4ecd197Z63kDXrhcF+rI6Xx0MSHWJ+4cUM6laoEgeYZEV7lhAh4CTDDLNBeK8H0E8wQAqxfWSOE8RXrWU4/xHTTzDz1gd7KgJuy1/6k3KWvmLee793w==</diagram></mxfile>
