+++ Sources/CoreDataCandy/Fetchable/Fetchable.swift
@@ -4,7 +4,7 @@
 
 import CoreData
 
-/// Can be fetched from the Core Data context. Adds simple to use fetch functions on a Core Data entity or `DatabaseModel`.
+/// Can be fetched from the Core Data context. Provide a way to specify the context where to fetch in a single place.
 public protocol Fetchable {
 
     /// If a non `nil` value is provided, all fetch requests will use this context
@@ -19,31 +19,8 @@ public extension Fetchable {
     static var context: NSManagedObjectContext? { nil }
 }
 
-/// A CoreData entity with augmented fetch requests
-public protocol FetchableEntity: NSManagedObject, DatabaseEntity, Fetchable {
-
-    /// The associated Database model name for fetching to be used when throwing a detailed error.
-    /// A default implementation is provided, which is useful when the entity is name accordingly
-    /// to its `DatabaseModel` counterpart with an 'Entity' suffix (e.g. 'PlayerEntity')
-    static var modelName: String { get }
-}
-
-public extension FetchableEntity where Self: NSManagedObject {
+public extension DatabaseEntity {
 
     /// `NSFetchRequest` to fetch the entity
-    static func newFetchRequest() -> NSFetchRequest<Self> { NSFetchRequest<Self>(entityName: String(describing: Self.self)) }
-}
-
-public extension FetchableEntity {
-
-    static var modelName: String {
-        // Default implementation of the model name
-        let entityIdentifier = "Entity"
-        var className = String(describing: self)
-
-        if className.hasSuffix(entityIdentifier) {
-            className.removeLast(entityIdentifier.count)
-        }
-        return className
-    }
+    static func newFetchRequest() -> NSFetchRequest<Self> { .init(entityName: String(describing: Self.self)) }
 }
