+++ Sources/CoreDataCandy/Fields&Relationships/Relationship/ChildrenInterfaceProtocol.swift
@@ -67,7 +67,7 @@ public extension ChildrenInterfaceProtocol where Entity: NSManagedObject, ChildM
     }
 }
 
-extension ChildrenInterfaceProtocol where Self: FieldPublisher, Self.Output == [ChildModel], ChildModel.Entity: FetchableEntity, Entity: NSManagedObject {
+extension ChildrenInterfaceProtocol where Self: FieldPublisher, Self.Output == [ChildModel], ChildModel.Entity: DatabaseEntity, Entity: NSManagedObject {
 
     func publisher(for entity: Entity, sortedBy sorts: [Sort<ChildModel.Entity>]) -> AnyPublisher<Output, Never> {
         publisher(for: entity)
