+++ Tests/CoreDataCandyTests/DatabaseModelTests.swift
@@ -85,7 +85,7 @@ final class DatabaseModelTests: XCTestCase {
 
 extension DatabaseModelTests {
 
-    final class StubEntity: NSManagedObject, FetchableEntity {
+    final class StubEntity: NSManagedObject, DatabaseEntity {
 
         static func fetchRequest() -> NSFetchRequest<StubEntity> {
             NSFetchRequest<StubEntity>(entityName: "Stub")
