+++ Tests/CoreDataCandyTests/RequestBuilderTests.swift
@@ -31,7 +31,7 @@ final class RequestBuilderTests: XCTestCase {
 
 extension RequestBuilderTests {
 
-    final class StubEntity: NSManagedObject, FetchableEntity {
+    final class StubEntity: NSManagedObject, DatabaseEntity {
 
         static func fetchRequest() -> NSFetchRequest<StubEntity> { NSFetchRequest<StubEntity>(entityName: "Stub") }
